Merging to release-4-lts: [TT-7808] In proxy-only mode, GraphQL query return errors from upstream (#5049) (#5058)
[TT-7808] In proxy-only mode, GraphQL query return errors from upstream
(#5049)

PR for [TT-7808](https://tyktech.atlassian.net/browse/TT-7808)

We pass the upstream headers and status codes to the client in
proxy-only mode. This is an existing feature. With this PR, we start
attaching the upstream's error messages to the response.

For example, say that a GraphQL service returns the following,
non-standard error message with HTTP 401 status code:

```json
{"message": "Unauthorized"}
```

We return the following error message with HTTP 401 message to the
client:

```json
{
    "errors": [
        {
            "message": "unable to resolve",
            "locations": [
                {
                    "line": 2,
                    "column": 5
                }
            ]
        }
    ],
    "data": null
}
```

It was not possible to reach out to the original error message from
upstream. With this PR, we started returning the following error message
with HTTP 401 status code:

```json
{
    "errors": [
        {
            "message": "unable to resolve",
            "locations": [
                {
                    "line": 2,
                    "column": 5
                }
            ],
            "extensions": {
                "message": "Unauthorized"
            }
        }
    ],
    "data": null
}
```

A new configuration item has been added to the API
definition(`graphql.proxy.return_errors_from_upstream`). This solution
is only valid for proxy-only mode and we only keep the original response
if upstream returns an error(status code >= 400).



[TT-7808]:

https://tyktech.atlassian.net/browse/TT-7808?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

[TT-7808]:
https://tyktech.atlassian.net/browse/TT-7808?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[TT-7808]:
https://tyktech.atlassian.net/browse/TT-7808?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[TT-7808]:
https://tyktech.atlassian.net/browse/TT-7808?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Burak Sezer <burak.sezer.developer@gmail.com>