Tt 851 process multivalue headers on response plugins (#5175)
<!-- Provide a general summary of your changes in the Title above -->

## Description

Updated the code in `gateway/coprocess.go` to work with headers
represented as Key => []Values
Is good to mention that now the object passed to the middleware contains
the old format header (one key - one value) and the new format header
(the mutivalue headers: one key- multiple values). So:

- If a new header is added in the old field, then this will be added as
well in the new field
- If a header is removed from the old field, then it will be removed as
well in the new field
- If a value of a header is updated then it will overwritte all the
values for that header in the new field.

In the end, the new field is the one being used to put all the headers
back in the response, because after calling the middleware a
synchronization process of headers is executed just to ensure that all
the changes are replicated in the new field, therefore, all the values
for every header are transmitted to the end user.

## Related Issue

TT-851

## Motivation and Context

<!-- Why is this change required? What problem does it solve? -->

## How This Has Been Tested

<!-- Please describe in detail how you tested your changes -->
<!-- Include details of your testing environment, and the tests -->
<!-- you ran to see how your change affects other areas of the code,
etc. -->
<!-- This information is helpful for reviewers and QA. -->

## Screenshots (if appropriate)

## Types of changes

<!-- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Refactoring or add test (improvements in base code or adds test
coverage to functionality)

## Checklist

<!-- Go over all the following points, and put an `x` in all the boxes
that apply -->
<!-- If there are no documentation updates required, mark the item as
checked. -->
<!-- Raise up any additional concerns not covered by the checklist. -->

- [ ] I ensured that the documentation is up to date
- [ ] I explained why this PR updates go.mod in detail with reasoning
why it's required
- [ ] I would like a code coverage CI quality gate exception and have
explained why