[TT-9397] removed the replace for gorm and inherit the version from tyk-pump (#5342)
<!-- Provide a general summary of your changes in the Title above -->

Removed the replace of gorm, this is being used for tests and currently
is forcing to use an older version of TykTechnologies/gorm, by remving
this, we inherit the version from the tyk-pump. In this way we avoid
issues in go plugins where gateway is using one version of gorm, but
pump another.

https://tyktech.atlassian.net/browse/TT-9397

Check ticket details

- Built the gw
- Ran tests

<!-- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Refactoring or add test (improvements in base code or adds test
coverage to functionality)

<!-- Go over all the following points, and put an `x` in all the boxes
that apply -->
<!-- If there are no documentation updates required, mark the item as
checked. -->
<!-- Raise up any additional concerns not covered by the checklist. -->

- [ ] I ensured that the documentation is up to date
- [ ] I explained why this PR updates go.mod in detail with reasoning
why it's required
- [ ] I would like a code coverage CI quality gate exception and have
explained why

(cherry picked from commit 108c9462bbe2ae0af698d30491db69f014e2aec6)