[TD-1809] tyk:r5-lts Once again templates are compatible (#5356)
Update/Sync Releng templates for tyk:release-5-lts

:robot: Beep, boop, this PR has been generated by
[automation](https://github.com/TykTechnologies/gromit/blob/master/cmd/policy.go)
at 2023-07-26 15:20:13.749922862 &#43;0800 &#43;08 m=&#43;32.132496907

This PR updates the release engineering specific code in the
release-5-lts branch of tyk, and has been generated because some releng
code was changed, and the corresponding templates were updated.

These changes should be limited to the `ci/` directory and `.github`
directory. This PR should be automatically merged with the branch
protection rules are satisfied.

If you think this was made by mistake, or if merging this might cause
any instabilities, please bring it up on Slack or log a jira in the TD
project.

## Repo Policy for tyk

Default Branch: **master**


### Active  Release Branches


| Branch | Releng Features | Source Branch | Go Version | Tests
| -------| -------------- | ------------- | ---------- | -----
| release-4-lts | el7,plugin-compiler | master | 1.15 | Go 1.15 Redis 6
| release-5.0.4 | el7,plugin-compiler | release-5-lts | 1.16 | Go 1.16
Redis 5,1.16,1.16-el7
| release-5-lts | el7,plugin-compiler | master | 1.16 | Go 1.16 Redis
5,1.16,1.16-el7
| release-4.0.14 | plugin-compiler,el7 | release-4-lts | 1.15 | Go 1.15
Redis 6
| release-5.1 |  | master |  | 
| release-5.1.1 |  | master |  | 


**Generated by gromit on Wed Jul 26 15:20:13 &#43;08 2023**

---------

Co-authored-by: Gromit <policy@gromit>
Co-authored-by: Asutosh <asutosh@tyk.io>