Merging to release-5.2.1: [TD-2193] Use correct tags for CI push and CI test (#5465)
[TD-2193] Use correct tags for CI push and CI test (#5465)

The old architecture of doing ci in a separate job made the larger
workflow seemingly more resilient but as the automation grows, failures
in CI will make the remainder of the release process unprocessable. So
doubling down on making CI push robust.

Removed ci/auto as it lives in tyk-ci now. Tested by re-using
[v5.2.1-alpha](https://github.com/TykTechnologies/tyk/actions/runs/5971645065).