[TT-10184] Support "detailed_tracing" in API def for GQL APIs (#5740)
PR for [TT-10184](https://tyktech.atlassian.net/browse/TT-10184)

I chose an interface-based design to implement the requested feature. In
this way, a minimum amount of intervention was required for the existing
implementation.

`GraphqlMiddleware Validation` span is also considered as a part of
"detailed tracing". So that span has been disabled if `detailed_tracing`
is set to `false`.

`TykOtelExecutorI` interface has been added to define Tyk/UDG-specific
methods.

If `detailed_tracing` is set to `false`:

![Screenshot 2023-11-06 at 13 51
39](https://github.com/TykTechnologies/tyk/assets/33702537/5a611578-15b7-4bce-9a25-7c1f3e54f6a5)

If `detailed_tracing` is set to `true`:

![Screenshot 2023-11-06 at 13 45
20](https://github.com/TykTechnologies/tyk/assets/33702537/6bb358a9-ba9a-4e3a-a59e-d95266275e61)



[TT-10184]:
https://tyktech.atlassian.net/browse/TT-10184?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ