[TT-10555] Fix/trimpath plugin compatibility and loading (#5784)
This PR changes the plugin compiler to adjust the module in use based on
$plugin_id argument.

Why this is needed:

Since we use `-trimpath` to increase plugin/gateway compatibility, the
filesystem paths are omitted from the built plugin. Previously we would
rely on that behaviour to reload plugins, by compiling the plugin in a
unique plugin path. Currently, the plugin build relies only on the go
module, so amending go.mod and the source file imports is needed.

This PR adds tests and test scaffolding to verify the changes.

Extended tests and behaviour:

- extends tyk plugin load to take multiple args, namely
plugin1.so,plugin2.so
- extended plugin compiler with debug output if DEBUG=1 env is provided
- implemented the two-plugin testcase using different plugin builds
- added complex-plugin test case with subpackages
- added plugin test case with invalid/simplistic go.mod module
- plugin compiler prints a warning if plugin module doesn't contain a
dot:

```
WARN: Plugin go.mod module doesn't contain a dot, consider amending it to prevent conflicts
      Current value: com
    Suggested value: github.com/org/plugin-repo
```

The replacement uses the go module name as a replacement pattern. If a
go module name like `net` would be set, the plugin compiler would
attempt replacing all imports with the prefix. That would conflict with
the standard library, `net/http` and other similar packages under the
namespace. It's also invalid for the plugin to declare it's module as
`plugin`, as a standard library package of the same name exists and may
conflict, result in unpredicted behaviour.

https://tyktech.atlassian.net/browse/TT-10555

---------

Co-authored-by: Tit Petric <tit@tyk.io>
(cherry picked from commit 3787c1383fe03cbb148a9dfa2fd749ee03698745)