[TT-10555] Reoder go.work/go.mod, add GO_TIDY env, test (#5822)
This PR reorders workspace creation to happen after we ensure that the
plugin has a go.mod file; This resolves a particular build issue due to
ordering:

```
INFO: Creating go.mod
go: creating new go.mod: module tyk.internal/tyk_plugin123
go: to add module requirements and sums:
	go mod tidy
directory . is contained in a module that is not one of the workspace modules listed in go.work. You can add the module to the workspace using go work use .
```

Now, if a plugin doesn't contain a go.mod file, a mod file is created in
the plugin, **before** the workspace is created. This ensures a passing
build. A test case has been added to verify (task
test:test-plugin-no-mod).

A debugging feature has been added, if GO_TIDY=1 exists in the
environment, `go mod tidy` would be run. It's not required by default
but it may aid to resolve some issues.

https://tyktech.atlassian.net/browse/TT-10555

Co-authored-by: Tit Petric <tit@tyk.io>