[SYSE-87] test-controller job decides test scope (#5943)
The complete list of features for _tyk:master_: releng.

:robot: Beep, boop, this PR has been generated by
[automation](https://github.com/TykTechnologies/gromit/blob/master/cmd/policy.go).
Changes should be limited to templated code. Where allowed, this PR will
be automatically merged when the branch protection rules are satisfied.
If you think this was made by mistake, or if merging this might cause
any instabilities, please bring it up on Slack or log a jira in the TD
project.

Co-authored-by: Gromit <policy@gromit>