[SYSE-87] r5-lts: test-controller job decides test scope (#5944)
The complete list of features for _tyk:release-5-lts_: releng, el7,
plugin-compiler-fix-vendor.

tyk:release-5-lts needs to support RHEL7. RHEL7 support was dropped in
5.1 and then brought back in 5.2.1. This PR may not directly affect
RHEL7 support but it may need testing on RHEL7 to make sure it still
works. Release/upgrade-tests will perform some basic tests on supported
OSes but this is no guarantee of everything being alright.
tyk:release-5-lts needs to be built _without_ module support and the
dependencies vendored for the sake of the plugin compiler. This is
implemented in release.yml. Note that the plugin compiler itself is
_not_ part of the templates and is built by its own workflow, this PR is
probably only about the gateway.
:robot: Beep, boop, this PR has been generated by
[automation](https://github.com/TykTechnologies/gromit/blob/master/cmd/policy.go).
Changes should be limited to templated code. Where allowed, this PR will
be automatically merged when the branch protection rules are satisfied.
If you think this was made by mistake, or if merging this might cause
any instabilities, please bring it up on Slack or log a jira in the TD
project.

Co-authored-by: Gromit <policy@gromit>