[TT-11091] Add support for custom ratelimit key (#5971)
<!-- Provide a general summary of your changes in the Title above -->

## Description
This PR introduces the ability to dynamically set the custom rate limit
key in the policies metadata. When `rate_limit_pattern` pattern is
specified in the policy metadata (e.g. "rate_limit_pattern":
"$tyk_meta.developer_id"), the ratelimit middleware calculates the
custom ratelimit key and uses it to enforce the ratelimit and quota.

## Related Issue
https://tyktech.atlassian.net/browse/TT-11091

<!-- This project only accepts pull requests related to open issues. -->
<!-- If suggesting a new feature or change, please discuss it in an
issue first. -->
<!-- If fixing a bug, there should be an issue describing it with steps
to reproduce. -->
<!-- OSS: Please link to the issue here. Tyk: please create/link the
JIRA ticket. -->

## Motivation and Context

For customers that expose APIs externally it’s quite important to have
flexibility on how to apply rate limit. For public APIs, it’s common to
apply rate limit not as a per credential but as per a developer or
organization profile because a commercial contract is create for
developer that includes all their credentials.

This also will help with abusing the free tier APIs where it’s currently
possible to reset a quota by rotating credentials. The problem is that
the rate limit key is the credential ID currently, so when a key or a
client is rotated, the new key ID is generated therefore effectively
reseting quota.

To improve the customer experience in both scenarios, we need to make
the rate limit key configurable.

## How This Has Been Tested
- Unit test covering the below scenarios
- Manual test (quota):
   - Create an auth token API
- Create a policy with that API, make sure that the quota is enabled
(for simplicity of testing - quota: 5 calls, never renewed), disable the
ratelimit to test the quota in isolation
- Specify the "rate_limit_pattern" in the policy metadata
("rate_limit_pattern": "$tyk_meta.developer_id")
   - Create two keys with that policy
   - In both key add the following metadata: developer_id = 123
   - Make 5 calls with the first key, all should be 200 OK
   - Make one more call with the first key, should be 403
   - Make one more call with the second key, should be 403 as well
- In one of the keys set developer_id in the metadata to another value
(e.g. 321) and save ***without quota reset**
   - Make one call with the that key, should be 200 OK
- Manual test (ratelimit):
   - Create an auth token API
- Create a policy with that API, make sure that the ratelimit is enabled
(for simplicity of testing - 5 calls per 3600 sec renewed), disable the
quota to test the ratelimit in isolation
- Specify the "rate_limit_pattern" in the policy metadata
("rate_limit_pattern": "$tyk_meta.developer_id")
   - Create two keys with that policy
   - In both key add the following metadata: developer_id = 123
   - Make 5 calls with the first key, all should be 200 OK
   - Make one more call with the first key, should be 429
   - Make one more call with the second key, should be 429 as well
- In one of the keys set developer_id in the metadata to another value
(e.g. 321) and save ***without quota reset**
   - Make one call with the that key, should be 200 OK

## Screenshots (if appropriate)

## Types of changes

<!-- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Refactoring or add test (improvements in base code or adds test
coverage to functionality)

## Checklist

<!-- Go over all the following points, and put an `x` in all the boxes
that apply -->
<!-- If there are no documentation updates required, mark the item as
checked. -->
<!-- Raise up any additional concerns not covered by the checklist. -->

- [ ] I ensured that the documentation is up to date
- [x] I explained why this PR updates go.mod in detail with reasoning
why it's required
- [ ] I would like a code coverage CI quality gate exception and have
explained why

---------

Co-authored-by: Sergey Lega <legabox@gmail.com>
Co-authored-by: Laurentiu Ghiur <laurentiu.ghiur@gmail.com>