refactor UDG and GraphQL poxy-only related code (TT-10533) (#5963)
This PR refactors all the code related to UDG and GraphQL proxy-only.

Those are the changes made:
- introduced a new internal package called `graphengine`. This will
contain all runtime code related to UDG and GraphQL
- introduced an interface `graphengine.Engine`. It servers as an
abstraction to the current ExecutionEngines that are stored in the API
spec. It now has implementation for the v1 engine
(`graphengine.EngineV1`) and v2 engine (`graphengine.EngineV2`)
- the internal package also contains a `graphqlGoToolsV1` struct that
serves as a stateless namespace struct to bundle all
`graphql-go-tools/v1` functionality. This way `graphql-go-tools/v1` and
`graphql-go-tools/v2` can co-exist in the internal package.
- removed a lot of code related to UDG and GraphQL proxy-only in the
`gateway` package. It should now be easier to understand the scope in
this package.

Not included in this PR:
- Callback functions like `OnBeforeStart` for WebSocket subscriptions
and `BeforeFetchHook` and `AfterFetchHook` are not yet moved to the
internal package. This will be done later.


## Types of changes

<!-- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [x] Refactoring or add test (improvements in base code or adds test
coverage to functionality)