Revert "[DX-1045]fix integer format issue  issue" (#6052)
## **User description**
Reverts TykTechnologies/tyk#6032

Apparently, `openapi-generator` creates `int32` when it doesn't
understand the format. This should cause issues with the gateway if you
have tests with negative values that you expect to pass. No one reported
such a case, however, since we are not sure if it's a breaking change or
not, and it's not worth the risk, we will revert this PR for now and
test it with a couple of tools to see what happens with existing tests.

Also, the HMAC skew is a "number" need to understand why.


___

## **Type**
enhancement


___

## **Description**
This PR updates the Swagger configuration to enhance data type
specifications and remove unnecessary constraints:
- Updated the `format` of JWT related fields
(`jwt_expires_at_validation_skew`, `jwt_issued_at_validation_skew`,
`jwt_not_before_validation_skew`) to `uint64` from `int64`.
- Removed `minimum` and `maximum` constraints for the JWT related
fields, making the configuration more flexible.
- Updated the `ssl_min_version` field `format` to `uint16` from `int64`
and removed its `minimum` and `maximum` constraints, simplifying the SSL
configuration.


___