[SYSE-336 release-5-lts] Followup from March template application (#6206)
Learning from the application:
Applying templates across all branches and repos runs up against the
usage limits. Larger runners are available for Team plans. This needs to
be explored.
A local run of release.yml is needed to test updates before pushing it
to all repos and waiting for >30mins to find out if it works.
For regular updates which collect disparate code changes, the git commit
log in the repos should reflect a summary of changes that went into it.
- [x] Mount go package and build cache into build container
- [x] nektos/act for local runs
- [x] dependabot alerts @devops for only .g/w/release.yml
- [x] r5-lts to test against pump and sink master
- [x] Larger runners for Actions
- [x] pk for byol build
- [x] Converge UI and API tests compose environments in tyk-ci

---------

Co-authored-by: Gromit <policy@gromit>