[SYSE-401 master] Model builds as a concept (#6728)
Hitherto, builds have been implemented using the features mechanism.
With the proliferation of builds, fips, s/390, ee and so on, this has
led to a confusing mess of conditionals in the templates.
With this change, builds are parameterised in the config file and follow
the same inheritance rules as the other parameters allowing for common
builds to be specified at upper levels.
A related change in policy match looks looks in tyk-ee instead of tyk.

---------

Co-authored-by: Gromit <policy@gromit>
Co-authored-by: Alok G Singh <alok@tyk.io>